Rename `admissionpluginsecret` to `shootkubeconfigsecretref`
This admission handler ensures that a kubeconfig is not removed from
Secrets in case it is referenced in shoots.

In the next commit, we augment it with verification for authorization
kubeconfigs.
