+++ skaffold.yaml
@@ -447,7 +447,6 @@ build:
             - pkg/admissioncontroller/seedidentity
             - pkg/admissioncontroller/webhook
             - pkg/admissioncontroller/webhook/admission
-            - pkg/admissioncontroller/webhook/admission/admissionpluginsecret
             - pkg/admissioncontroller/webhook/admission/auditpolicy
             - pkg/admissioncontroller/webhook/admission/authenticationconfig
             - pkg/admissioncontroller/webhook/admission/authorizationconfig
@@ -456,6 +455,7 @@ build:
             - pkg/admissioncontroller/webhook/admission/namespacedeletion
             - pkg/admissioncontroller/webhook/admission/resourcesize
             - pkg/admissioncontroller/webhook/admission/seedrestriction
+            - pkg/admissioncontroller/webhook/admission/shootkubeconfigsecretref
             - pkg/admissioncontroller/webhook/auth/seed
             - pkg/admissioncontroller/webhook/auth/seed/graph
             - pkg/apis/core
